+++ src/feature/auth/ui/forgot-password-form/index.tsx
@@ -13,16 +13,16 @@ import {
   forgotPasswordSchema,
 } from '@/feature/auth/model/utils/validators'
 import {
+  AuthRoutes,
   Button,
   Card,
   ControlledInput,
-  Recaptcha,
-  Text,
-  ReturnComponent,
-  UseFormRef,
-  AuthRoutes,
   EMPTY_STRING,
+  Recaptcha,
+  ReturnComponent,
   SM_BREAKPOINT,
+  Text,
+  UseFormRef,
   useFormRevalidateWithLocale,
   useResponsive,
   useTranslation,
