+++ src/shared/ui/date-picker/date-picker.tsx
@@ -1,10 +1,8 @@
 import * as React from 'react'
 import { ChangeEvent, forwardRef, useId } from 'react'
 
+import { Button, Popover, PopoverContent, PopoverTrigger, ReturnComponent, cn } from '@/shared'
 import { CalendarIcon, CalendarOutlineIcon } from '@/shared/assets/icons'
-import { cn } from '@/shared/lib'
-import { ReturnComponent } from '@/shared/types'
-import { Button, Popover, PopoverContent, PopoverTrigger } from '@/shared/ui'
 
 import { Calendar, CalendarProps } from './calendar'
