+++ src/shared/ui/text/text.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { CSSProperties, ComponentPropsWithoutRef, ElementType, ReactNode } from 'react'
 
-import { ReturnComponent } from '@/shared/types'
+import { ReturnComponent } from '@/shared'
 import { clsx } from 'clsx'
 
 export type TextColor = 'dark' | 'error' | 'info' | 'lightDark' | 'primary' | 'success' | 'warning'
