+++ src/widgets/header/ui/auth-buttons/auth-buttons.tsx
@@ -1,9 +1,6 @@
 import * as React from 'react'
 
-import { AuthRoutes } from '@/shared/constants'
-import { useTranslation } from '@/shared/lib'
-import { ReturnComponent } from '@/shared/types'
-import { Button } from '@/shared/ui'
+import { AuthRoutes, Button, ReturnComponent, useTranslation } from '@/shared'
 import Link from 'next/link'
 
 export const AuthButtons = (): ReturnComponent => {
